requirements.txt
Pillow==7.1.2
responder==2.0.5
scikit-learn==0.23.1